move GraphBaseEntityService to package service.base
