+++ src/main/java/org/sbml4j/controller/NetworksApiController.java
@@ -43,10 +43,10 @@ import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
 import org.sbml4j.model.warehouse.MappingNode;
 import org.sbml4j.service.ConfigService;
 import org.sbml4j.service.FlatSpeciesService;
-import org.sbml4j.service.GraphBaseEntityService;
 import org.sbml4j.service.MyDrugService;
 import org.sbml4j.service.ProvenanceGraphService;
 import org.sbml4j.service.WarehouseGraphService;
+import org.sbml4j.service.base.GraphBaseEntityService;
 import org.sbml4j.service.networks.NetworkResourceService;
 import org.sbml4j.service.networks.NetworkService;
 import org.sbml4j.service.utility.FileCheckService;
