+++ src/main/java/org/sbml4j/controller/SbmlApiController.java
@@ -39,12 +39,12 @@ import org.sbml4j.model.warehouse.FileNode;
 import org.sbml4j.model.warehouse.Organism;
 import org.sbml4j.model.warehouse.PathwayNode;
 import org.sbml4j.service.ConfigService;
-import org.sbml4j.service.GraphBaseEntityService;
 import org.sbml4j.service.PathwayService;
 import org.sbml4j.service.ProvenanceGraphService;
 import org.sbml4j.service.SBMLSimpleModelService;
 import org.sbml4j.service.UtilityService;
 import org.sbml4j.service.WarehouseGraphService;
+import org.sbml4j.service.base.GraphBaseEntityService;
 import org.sbml4j.service.utility.FileCheckService;
 import org.sbml4j.service.warehouse.DatabaseNodeService;
 import org.sbml4j.service.warehouse.FileNodeService;
