+++ src/main/java/org/sbml4j/service/networks/NetworkService.java
@@ -50,11 +50,11 @@ import org.sbml4j.service.ContextService;
 import org.sbml4j.service.CsvService;
 import org.sbml4j.service.FlatEdgeService;
 import org.sbml4j.service.FlatSpeciesService;
-import org.sbml4j.service.GraphBaseEntityService;
 import org.sbml4j.service.ProvenanceGraphService;
 import org.sbml4j.service.UtilityService;
 import org.sbml4j.service.WarehouseGraphService;
 import org.sbml4j.service.SimpleSBML.SBMLSpeciesService;
+import org.sbml4j.service.base.GraphBaseEntityService;
 import org.sbml4j.service.warehouse.MappingNodeService;
 import org.sbml4j.service.warehouse.OrganismService;
 import org.slf4j.Logger;
