+++ src/main/java/org/sbml4j/service/warehouse/DatabaseNodeService.java
@@ -18,7 +18,7 @@ import java.util.List;
 import org.sbml4j.model.warehouse.DatabaseNode;
 import org.sbml4j.model.warehouse.Organism;
 import org.sbml4j.repository.warehouse.DatabaseNodeRepository;
-import org.sbml4j.service.GraphBaseEntityService;
+import org.sbml4j.service.base.GraphBaseEntityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
