+++ src/main/java/org/sbml4j/service/warehouse/PathwayCollectionNodeService.java
@@ -26,10 +26,10 @@ import org.sbml4j.model.warehouse.Organism;
 import org.sbml4j.model.warehouse.PathwayCollectionNode;
 import org.sbml4j.model.warehouse.PathwayNode;
 import org.sbml4j.repository.warehouse.PathwayCollectionNodeRepository;
-import org.sbml4j.service.GraphBaseEntityService;
 import org.sbml4j.service.PathwayService;
 import org.sbml4j.service.ProvenanceGraphService;
 import org.sbml4j.service.WarehouseGraphService;
+import org.sbml4j.service.base.GraphBaseEntityService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
